+++ tests/test_admin/test_forms.py
@@ -25,8 +25,8 @@ from enterprise.admin.forms import (
 )
 from enterprise.admin.utils import ValidationMessages
 from enterprise.constants import ENTERPRISE_ADMIN_ROLE
-from enterprise.models import EnterpriseCustomer, SystemWideEnterpriseRole
-from test_utils import fake_enrollment_api
+from enterprise.models import EnterpriseCustomer, EnterpriseCustomerReportingConfiguration, SystemWideEnterpriseRole
+from test_utils import TEST_PGP_KEY, fake_enrollment_api
 from test_utils.factories import (
     AdminNotificationFactory,
     EnterpriseCatalogQueryFactory,
@@ -758,6 +758,53 @@ class TestEnterpriseCustomerReportingConfigAdminForm(unittest.TestCase):
             data=self.form_data,
         )
         assert not form.is_valid()
+
+    def test_invalid_pgp_key(self):
+        """
+        Clean should throw errors about invalid pgp keys
+        """
+        self.form_data['pgp_encryption_key'] = 'invalid-pgp-key'
+        form = EnterpriseCustomerReportingConfigAdminForm(
+            data=self.form_data,
+        )
+        assert not form.is_valid()
+
+    def test_valid_pgp_key(self):
+        """
+        Clean should accept valid pgp keys.
+
+        Also, user can choose to not provide a PGP key. Error should not be raised in this case.
+        """
+        self.form_data['pgp_encryption_key'] = TEST_PGP_KEY
+        form = EnterpriseCustomerReportingConfigAdminForm(
+            data=self.form_data,
+        )
+        assert form.is_valid()
+
+        # Empty value for PGP key is allowed
+        self.form_data['pgp_encryption_key'] = ''
+        form = EnterpriseCustomerReportingConfigAdminForm(
+            data=self.form_data,
+        )
+        assert form.is_valid()
+
+    def test_validate_manual_reporting(self):
+        """
+        Verify that manual reporting validation works as exepcted.
+
+        Only enterprises enabled for manual reporting can create reporting configuration.
+        """
+        report_config_data = dict(self.form_data, data_type=EnterpriseCustomerReportingConfiguration.DATA_TYPE_GRADE)
+        form = EnterpriseCustomerReportingConfigAdminForm(
+            data=report_config_data,
+        )
+        assert not form.is_valid()
+        errors = form.errors
+        message = '"{data_type}" data type is not supported for enterprise customer "{enterprise_customer.name}". Please select a different data type.'.format(  # pylint: disable=line-too-long
+            enterprise_customer=self.ent_customer1,
+            data_type=EnterpriseCustomerReportingConfiguration.DATA_TYPE_GRADE,
+        )
+        assert errors == {'data_type': [message]}
 
 
 @ddt.ddt