+++ CHANGELOG.rst
@@ -17,6 +17,22 @@ Unreleased
 ----------
 * Nothing
 
+[3.58.12]
+---------
+fix: make integrated channels api response record field readonly in admin view
+
+[3.58.11]
+---------
+fix: Added model field validation for pgp_encryption_key
+
+[3.58.10]
+---------
+refactor: update enterprise manual reporting to make it generic
+
+[3.58.9]
+--------
+fix: Added validation for PGP encryption key entered by the user.
+
 [3.58.8]
 --------
 feat: integration sync reporting viewset improvements